# user_service/docs/swagger.yaml
info:
  contact: {}
paths:
  /api/users:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: ""
      summary: Delete user by id endpoint
      tags:
      - Users
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
      summary: Get all users endpoint
      tags:
      - Users
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: ""
      summary: Partially update user endpoint
      tags:
      - Users
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: ""
      summary: Create user endpoint
      tags:
      - Users
  /api/users/auth:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
      summary: Get user by username and password. Needs for authorization
      tags:
      - Users
  /api/users/id/:id:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
      summary: Get user by user id
      tags:
      - Users
  /api/users/tickets:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: ""
      summary: Delete prize by user id and prize id
      tags:
      - Tickets
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: ""
      summary: Add prize by user id and prize id
      tags:
      - Tickets
  /api/users/tickets/free/:id:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
      summary: Get prize status
      tags:
      - Tickets
  /api/users/username/:username:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
      summary: Get user by username endpoint
      tags:
      - Users
  /userapi/heartbeat:
    get:
      responses:
        "204":
          description: ""
        "400":
          description: ""
      summary: Heartbeat metric
      tags:
      - Metrics
swagger: "2.0"